Move more epoch code into one file
